feat: adicionar operações CRUD para consultas e treinos, melhorar controle de sessão e aprimorar modais da interface
- Implementadas rotas PUT e DELETE para consultas em ConsultaRoutes.js.
- Adicionadas rotas GET para buscar detalhes de treino e listar treinos em TreinoRoutes.js.
- Melhorado o controle de sessão em controlarSessao.js para restringir o acesso com base nos papéis de usuário.
- Introduzidos alertas modais para feedback nas páginas de consulta e dashboards.
- Atualizadas as páginas de consulta e treino para utilizar modais em alertas e confirmações.
- Melhorado o estilo das páginas de dashboard e treino para uma melhor experiência do usuário.
- Refatoradas funções JavaScript para lidar com carregamento e exibição de dados com mais eficiência.

+++ frontend/pages/dashboards/indexPersonal.js
@@ -98,7 +98,8 @@ async function carregarTreinos(id, card) {
   listaTreinos.innerHTML = `<p class="cards-div-message">Carregando treinos...</p>`;
 
   try {
-    const responseTreinos = await fetch(`http://localhost:3000/treino?id=` + id, {
+    const responseTreinos = await fetch(`http://localhost:3000/treino?id=${id}`, {
+      method: "GET",
       credentials: "include"
     });
     const dataTreinos = await responseTreinos.json();
@@ -109,7 +110,7 @@ async function carregarTreinos(id, card) {
     }
 
     const treinos = Array.isArray(dataTreinos.sucesso) ? dataTreinos.sucesso : [];
-    
+
     listaTreinos.innerHTML = `
       <a class="card-botoes" href="../treino/index.html?id_cliente=${id}">Criar Treino</a>
     `;
@@ -119,13 +120,13 @@ async function carregarTreinos(id, card) {
       return;
     }
 
-    listaTreinos.innerHTML += treinos.map((treino) => `
+    listaTreinos.innerHTML += treinos.map((t) => `
     <div class="card-cliente card-dieta">
       <div class="card-content">
-        <div class="card-texto"><span>Título:</span> ${treino.titulo_treino}</div>
+        <div class="card-texto"><span>Título:</span> ${t.nome_treino}</div>
         <div class="card-botoes-container dieta-acoes">
-          <div class="card-botoes" data-treino-id="${treino.id_treino}" onclick="excluirTreino(${treino.id_treino}, ${id})">Excluir Treino</div>
-          <a class="card-botoes" data-treino-id="${treino.id_treino}" href="../treino/index.html?id_treino=${treino.id_treino}">Atualizar Treino</a>
+          <div class="card-botoes" data-treino-id="${t.id_treino}" onclick="excluirTreino(${t.id_treino}, ${id})">Excluir Treino</div>
+          <a class="card-botoes" data-treino-id="${t.id_treino}" href="../treino/index.html?id_treino=${t.id_treino}">Atualizar Treino</a>
         </div>
       </div>
     </div>
@@ -162,20 +163,39 @@ async function verDados(id) {
   }
 }
 
-async function excluirTreino(treinoId, clienteId) {
+function excluirTreino(treinoId, clienteId) {
   const card = document.getElementById(clienteId);
-  try {
-    const response = await fetch("http://localhost:3000/treino/" + treinoId, {
-      method: "DELETE",
-      credentials: "include"
-    });
-    const data = await response.json();
-    if (!data.sucesso) return;
-    await carregarTreinos(clienteId, card);
-    alert(data.sucesso);
-  } catch (error) {
-    alert("Erro ao deletar treino");
-  }
+  abrirModal("Deletar Treino", "Tem certeza que deseja deletar esse treino?");
+  const aceitarBtn = document.getElementById("modalAcceptBtn");
+  if (!aceitarBtn) return;
+  aceitarBtn.addEventListener("click", async (e) => {
+    e.preventDefault();
+    try {
+      const response = await fetch("http://localhost:3000/treino/" + treinoId, {
+        method: "DELETE",
+        credentials: "include"
+      });
+      const data = await response.json();
+      if (!data.sucesso) return;
+      await carregarTreinos(clienteId, card);
+      fecharModal();
+    } catch (error) {
+      fecharModal();
+      abrirAlerta("Erro", "Erro ao deletar treino");
+    }
+  });
+}
+function abrirAlerta(titulo, descricao) {
+  document.querySelector(".modal").classList.add("is-active");
+  document.querySelector(".modal").innerHTML = `
+	<div class="modal-content">
+      <h2>${titulo}</h2>
+      <p>${descricao}</p>
+      <div>
+        <button onclick="fecharModal()">Fechar</button>
+      </div>
+    </div>
+	`;
 }
 
 function fecharModal() {
